feat: make crud to PLAN Class

+++ api/models/plan.py
@@ -1,5 +1,4 @@
-from sqlalchemy import Column, String, Integer, DECIMAL
-
+from sqlalchemy import Column, String, Integer, Float
 from api.core.database import Base
 
 
@@ -7,6 +6,6 @@ class Plan(Base):
     __tablename__ = "plans"
     id: Integer = Column(Integer, primary_key=True, autoincrement=True)
     name: String = Column(String, unique=True)
-    price: DECIMAL = Column(DECIMAL, default=0.0)
+    price: Float = Column(Float, default=0.0)
     user_limit: Integer = Column(Integer, default=0)
     scheduling_limit: Integer = Column(Integer, default=0)
